+++ src/app/admin/admin.module.ts
@@ -26,6 +26,7 @@ import {SubmissionEffects} from "./submission/submission.effects";
 import {TestCaseEffects} from "./test-case/test-case.effects";
 import { RoleComponent } from './role/role.component';
 import {RoleViewComponent} from "./role/role-view/role-view.component";
+import {RoleEditComponent} from "./role/role-edit/role-edit.component";
 import {RoleEffects} from "./role/role.effects";
 import {PermissionComponent} from "./permission/permission.component";
 import {PermissionEffects} from "./permission/permission.effects";
@@ -44,6 +45,7 @@ import {PermissionEffects} from "./permission/permission.effects";
     ProblemTestCaseEditComponent,
     RoleComponent,
     RoleViewComponent,
+    RoleEditComponent,
     PermissionComponent
   ],
   imports: [